list redeemd preznts for user
